+++ config.sh
@@ -44,3 +44,24 @@ CERT_TEST=
 
 # Use --force-renewal option. (true or false) Default: false
 CERT_FORCE_RENEWAL=
+
+# Docker images
+IMAGE_KEYROCK=fiware/idm:8.0.0
+IMAGE_WILMA=fiware/pep-proxy:8.0.0
+IMAGE_ORION=fiware/orion:3.2.0
+IMAGE_CYGNUS=fiware/cygnus-ngsi:2.10.0
+IMAGE_COMET=fiware/sth-comet:2.8.0
+IMAGE_WIRECLOUD=fiware/wirecloud:1.3.1
+IMAGE_NGSIPROXY=fiware/ngsiproxy:1.2.2
+IMAGE_QUANTUMLEAP=orchestracities/quantumleap:0.8.1
+
+IMAGE_MONGO=mongo:4.4
+IMAGE_MYSQL=mysql:5.7
+IMAGE_POSTGRES=postgres:11
+IMAGE_CRATE=crate:4.1.4
+
+IMAGE_NGINX=nginx:1.21
+IMAGE_REDIS=redis:6
+IMAGE_ELASTICSEARCH=elasticsearch:2.4
+IMAGE_MEMCACHED=memcached:1
+IMAGE_GRAFANA=grafana/grafana:6.1.6