+++ setup/clean.sh
@@ -4,7 +4,7 @@ if [ -d ./data ]; then
   sudo rm -fr data
 fi
 
-for file in docker-compose.yml docker-cert.yml docker-keyrock.yml
+for file in docker-compose.yml docker-cert.yml docker-idm.yml
 do
   if [ -e ${file} ]; then
     sudo /usr/local/bin/docker-compose -f ${file} down