+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## FIWARE Big Bang v0.3.0-next
+
 ## FIWARE Big Bang v0.3.0 - 25 Sep, 2021
 
 -   Fix certbot image (#55)